Seperating the GUI code from the core.
